requirements.txt
streamlit==1.44.1
reportlab==4.3.1
pillow==11.1.0
pypdf==5.3.1
